Introducing Schema Crawler (#25)
* Schema crawler integration and database abstraction
* schema name should be uppercase for Oracle
* MS SQL scripts removed
* Hard-coded SQL quotes removed, no more useless quoting. No more database specific data types
* Hard-coded SQL quotes removed, no more useless quoting. No more database specific data types
* Hard-coded SQL quotes removed, no more useless quoting. No more database specific data types
* No driver specification needed
* Back compatibility for JDBC driver loading, column exclusion by type, inverted tables exclusion fixes #14
* fixed wrong data type

+++ neo4j-etl/src/test/java/org/neo4j/etl/neo4j/importcsv/fields/DataTest.java
@@ -13,7 +13,7 @@ public class DataTest
         try
         {
             // when
-            new Data( null, Neo4jDataType.Int );
+            new Data( null, Neo4jDataType.Long );
             fail( "Expected NullPointerException" );
         }
         catch ( NullPointerException e )
@@ -29,7 +29,7 @@ public class DataTest
         try
         {
             // when
-            new Data( null, Neo4jDataType.Int, true );
+            new Data( null, Neo4jDataType.Long, true );
             fail( "Expected NullPointerException" );
         }
         catch ( NullPointerException e )
@@ -45,7 +45,7 @@ public class DataTest
         try
         {
             // when
-            new Data( " \t", Neo4jDataType.Int );
+            new Data( " \t", Neo4jDataType.Long );
             fail( "Expected IllegalArgumentException" );
         }
         catch ( IllegalArgumentException e )
@@ -61,7 +61,7 @@ public class DataTest
         try
         {
             // when
-            new Data( " \t", Neo4jDataType.Int, true );
+            new Data( " \t", Neo4jDataType.Long, true );
             fail( "Expected IllegalArgumentException" );
         }
         catch ( IllegalArgumentException e )
